Build without warnings in the default configuration
A clean build reported 1321 warnings, all of them pre-existing and none of
them fatal, which made the output useless for spotting a new one.

Most were -Wreorder: nine constructors listed their members in an order that
did not match the declaration order. Reordering the lists changes nothing at
runtime, since members are initialised in declaration order either way, which
is exactly why the warning is worth heeding - the written order was describing
something that never happened.

One was a real defect. Log::init_spdlog_logging assigned the result of
std::string::find to an int, so npos was truncated and only compared equal to
npos again through the way the value converts back. It happened to work here;
it now uses the right type.

The two -Wswitch sites are treated differently on purpose. MqttClient::start
used a one-armed switch as a re-entry guard, which reads better as an if.
onDisconnect really does dispatch on state, so the missing state gets an arm
that says there is nothing to do. Neither gets a default label: that would
silence the only compile time check for a forgotten arm, as mqttpayload.cpp
already notes.

The misleading indentation in ConverterArgParser::setArgValue was not a bug,
because the unbraced if throws and the mis-indented line cannot be reached
when it fires, but the layout implied otherwise and now has braces. The unused
variables were all genuinely dead, including one in exprconv_float_tests that
GCC cannot report because std::string has a non-trivial destructor, and a
commented-out mosquitto callback naming a method that exists nowhere.

+++ unittests/mockedmodbuscontext.cpp
@@ -190,13 +190,13 @@ MockedModbusContext::Slave::hasError(const std::map<int, MockedModbusContext::Re
 
 const modmqttd::RegisterPoll&
 MockedModbusContext::Slave::getIssuedReadCall(int number) const {
-    REQUIRE(mIssuedReadCalls.size() > number);
+    REQUIRE(static_cast<int>(mIssuedReadCalls.size()) > number);
     return mIssuedReadCalls[number];
 }
 
 const modmqttd::RegisterWrite&
 MockedModbusContext::Slave::getIssuedWriteCall(int number) const {
-    REQUIRE(mIssuedWriteCalls.size() > number);
+    REQUIRE(static_cast<int>(mIssuedWriteCalls.size()) > number);
     return mIssuedWriteCalls[number];
 }
